PA: Scrape people with pupa.
